refactor: remove deprecated jx create cluster
